feat: storybook에 globalStyle과 theme 적용 (#3)
